+++ types/ingest/index.ts
@@ -1,5 +1,6 @@
 export * from './CronsSectionProps'
 export * from './EmailPipelineCardProps'
+export * from './FullRescanDialogProps'
 export * from './PollProgressListProps'
 export * from './PollStatusResponseData'
 export * from './ProcessedEmail'