+++ utils/install/index.ts
@@ -5,6 +5,7 @@ export { DEFAULT_IMAP_PORT } from './defaultImapPort'
 export { DEFAULT_INTERVAL } from './defaultInterval'
 export { installReducer } from './installReducer'
 export { MIN_SECRET_LENGTH } from './minSecretLength'
+export { parseIngestionDaysBack } from './parseIngestionDaysBack'
 export { timingSafeStringEqual } from './timingSafeStringEqual'
 
 export * from '../../validators/install/validateInstallBody'